Starting test refactoring
