+++ java/pom.xml
@@ -93,6 +93,12 @@
             <artifactId>jackson-databind</artifactId>
             <version>2.12.7.1</version>
         </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>1.18.34</version>
+            <scope>provided</scope>
+        </dependency>
 
         <dependency>
             <groupId>junit</groupId>
