+++ java/src/main/java/org/opensky/enums/PositionSource.java
@@ -1,0 +1,4 @@
+package org.opensky.enums;
+
+public enum PositionSource {
+}
