+++ java/src/main/java/org/opensky/enums/RequestType.java
@@ -1,0 +1,6 @@
+package org.opensky.enums;
+
+public enum RequestType {
+    GET_STATES,
+    GET_MY_STATES
+}
